Crude way to work in the new 3.5GPT-turbo end point
This is a exetremely rough hack to make it work so I can move on with a project. Ultimately you will need a new settings option to toggle the endpoint to include 'chat'/completions and back to the old /completions endpoint.
Just need to stay on davinci in the editor and not use any other setting. The other models won't work. Only 3.5 on the Davinci setting